requirements.txt
flask
face_recognition
requests
google-api-python-client
pillow
gunicorn
